WinMME: Fix a possible bug with outgoing MIDI timing.  It might not actually be a bug.  Either way, it's a better way to convert from jack_time_t to LONGLONG.

+++ windows/winmme/JackWinMMEOutputPort.cpp
@@ -145,7 +145,8 @@ JackWinMMEOutputPort::Execute()
             LARGE_INTEGER due_time;
 
             // 100 ns resolution
-            due_time.QuadPart = - ((frame_time - current_time) * 10);
+            due_time.QuadPart =
+                -((LONG LONG) ((frame_time - current_time) * 10));
             if (! SetWaitableTimer(timer, &due_time, 0, NULL, NULL, 0)) {
                 WriteOSError("JackWinMMEOutputPort::Execute",
                              "ChangeTimerQueueTimer");
